+++ app/src/main/java/com/adsamcik/starlitcoffee/navigation/Routes.kt
@@ -52,6 +52,12 @@ object BarcodeScanner
 object GuidedScan
 
 @Serializable
+data class GuidedScanDraft(val sessionId: String)
+
+@Serializable
+data class BagDraft(val sessionId: String)
+
+@Serializable
 object OnboardingMethods
 
 @Serializable