+++ app/src/main/java/com/adsamcik/starlitcoffee/ui/component/SuggestingTextField.kt
@@ -2,6 +2,7 @@ package com.adsamcik.starlitcoffee.ui.component
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExposedDropdownMenu
 import androidx.compose.material3.ExposedDropdownMenuBox
@@ -110,7 +111,7 @@ fun SuggestingTextField(
                 },
         )
 
-        ExposedDropdownMenu(
+        DropdownMenu(
             expanded = showSuggestions,
             onDismissRequest = { menuExpanded = false },
         ) {